Move the last query code to Plugin
