+++ neuron/src/lib/Neuron/Frontend/Impulse.hs
@@ -23,7 +23,6 @@ import Data.TagTree (Tag (..))
 import Data.Tagged
 import qualified Data.Text as T
 import Data.Tree (Tree (..))
-import qualified Neuron.Frontend.Query.View as QueryView
 import Neuron.Frontend.Route (NeuronWebT)
 import Neuron.Frontend.Route.Data.Types
 import qualified Neuron.Frontend.Theme as Theme
@@ -31,6 +30,7 @@ import Neuron.Frontend.Widget (LoadableData, divClassVisible, elVisible)
 import qualified Neuron.Frontend.Widget as W
 import Neuron.Frontend.Zettel.View (renderZettelParseError)
 import qualified Neuron.Frontend.Zettel.View as ZettelView
+import qualified Neuron.Plugin.Plugins.Links as Links
 import qualified Neuron.Plugin.Plugins.Tags as Tags
 import Neuron.Zettelkasten.ID (ZettelID (..))
 import Neuron.Zettelkasten.Zettel
@@ -185,7 +185,7 @@ renderErrors issues = do
                               "Links in vain to: "
                                 <> T.intercalate ", " (toList $ unZettelID . untag <$> missingZids)
                         elAttr "span" ("title" =: tooltip) $ do
-                          QueryView.renderZettelLinkIDOnly zid slug
+                          Links.renderZettelLinkIDOnly zid slug
   where
     renderError zid zError = do
       case zError of
@@ -208,7 +208,7 @@ renderErrors issues = do
     errorMessageHeader zid = \case
       ZettelError_ParseError (slug, _) -> do
         text "Zettel "
-        QueryView.renderZettelLinkIDOnly zid slug
+        Links.renderZettelLinkIDOnly zid slug
         text " failed to parse"
       ZettelError_AmbiguousID _files -> do
         text $
@@ -246,7 +246,7 @@ renderForest treesDyn = do
 zettelLink :: (DomBuilder t m, PostBuild t m) => Zettel -> NeuronWebT t m () -> NeuronWebT t m ()
 zettelLink z w = do
   el "li" $ do
-    QueryView.renderZettelLink Nothing Nothing def z
+    Links.renderZettelLink Nothing Nothing def z
     w
 
 searchInput ::
