+++ neuron/src/lib/Neuron/Plugin/Plugins/Links.hs
@@ -16,6 +16,8 @@ module Neuron.Plugin.Plugins.Links
   ( plugin,
     routePluginData,
     renderHandleLink,
+    renderZettelLink,
+    renderZettelLinkIDOnly,
     renderPanel,
     preJsonStrip,
   )
@@ -30,11 +32,12 @@ import Control.Monad.Writer
 import qualified Data.Dependent.Map as DMap
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
+import Data.Some
 import Data.Tagged (Tagged (Tagged), untag)
 import qualified Data.Text as T
-import qualified Neuron.Frontend.Query.View as Q
-import Neuron.Frontend.Route (NeuronWebT)
+import Neuron.Frontend.Route (NeuronWebT, Route (Route_Zettel), neuronRouteLink)
 import Neuron.Frontend.Route.Data.Types (LinksData (..))
+import Neuron.Frontend.Widget (elNoSnippetSpan, elTime)
 import Neuron.Plugin.Type (Plugin (..))
 import Neuron.Zettelkasten.Connection
   ( Connection (Folgezettel, OrdinaryConnection),
@@ -42,7 +45,7 @@ import Neuron.Zettelkasten.Connection
   )
 import qualified Neuron.Zettelkasten.Graph as G
 import Neuron.Zettelkasten.Graph.Type (ZettelGraph)
-import Neuron.Zettelkasten.ID (ZettelID, getZettelID)
+import Neuron.Zettelkasten.ID (Slug, ZettelID (unZettelID), getZettelID)
 import Neuron.Zettelkasten.Zettel
 import qualified Neuron.Zettelkasten.Zettel as Z
 import Reflex.Dom.Core hiding (count, mapMaybe, tag)
@@ -99,7 +102,7 @@ renderPanel elNeuronPandoc LinksData {..} = do
       elClass "ul" "backlinks" $ do
         forM_ links $ \((conn, ctxList), zl) ->
           el "li" $ do
-            Q.renderZettelLink Nothing (Just conn) def zl
+            renderZettelLink Nothing (Just conn) def zl
             elAttr "ul" ("class" =: "context-list" <> "style" =: "zoom: 85%;") $ do
               forM_ ctxList $ \ctx -> do
                 elClass "li" "item" $ do
@@ -187,9 +190,83 @@ renderZettelLinkMay ::
   (PandocBuilder t m, PostBuild t m) => Maybe [Inline] -> Either MissingZettel (Connection, Zettel) -> NeuronWebT t m ()
 renderZettelLinkMay minner = \case
   Left (untag -> zid) ->
-    Q.renderMissingZettelLink zid
+    renderMissingZettelLink zid
   Right (conn, target) ->
-    Q.renderZettelLink (elPandocInlines <$> minner) (Just conn) Nothing target
+    renderZettelLink (elPandocInlines <$> minner) (Just conn) Nothing target
+
+-- | Render a link to an individual zettel.
+renderZettelLink ::
+  (DomBuilder t m, PostBuild t m) =>
+  -- | Link inner text
+  Maybe (m ()) ->
+  -- | Connection type to display
+  Maybe Connection ->
+  -- | Link theme
+  Maybe LinkView ->
+  Zettel ->
+  NeuronWebT t m ()
+renderZettelLink mInner conn (fromMaybe def -> linkView) Zettel {..} = do
+  let connClass = show <$> conn
+      rawClass = const (Just "errors") =<< untag zettelContent
+      mextra =
+        case linkView of
+          LinkView_Default ->
+            Nothing
+          LinkView_ShowDate ->
+            elTime <$> zettelDate
+          LinkView_ShowID ->
+            Just $ el "tt" $ text $ unZettelID zettelID
+      classes :: [Text] = catMaybes $ [Just "zettel-link-container"] <> [connClass, rawClass]
+  elClass "span" (T.intercalate " " classes) $ do
+    forM_ mextra $ \extra ->
+      elNoSnippetSpan ("class" =: "extra monoFont") $ do
+        extra
+        -- The extra space is so that double clicking on this extra text
+        -- doesn't select the title next.
+        text " "
+    elAttr "span" ("class" =: "zettel-link" <> maybe mempty ("title" =:) linkTooltip) $ do
+      let linkInnerHtml = fromMaybe (text zettelTitle) mInner
+      neuronRouteLink (Some $ Route_Zettel zettelSlug) mempty linkInnerHtml
+      elConnSuffix conn
+  where
+    -- If there is custom inner text, put zettel title in tooltip.
+    linkTooltip
+      | isJust mInner = Just $ "Zettel: " <> zettelTitle
+      | otherwise = Nothing
+    elConnSuffix :: DomBuilder t m => Maybe Connection -> m ()
+    elConnSuffix mconn =
+      case mconn of
+        Just Folgezettel -> elNoSnippetSpan mempty $ do
+          elAttr "sup" ("title" =: "Branching link (folgezettel)") $ text "ᛦ"
+        _ -> pure mempty
+
+-- TODO: Eventually refactor this function to reuse what's in renderZettelLink
+renderMissingZettelLink :: DomBuilder t m => ZettelID -> m ()
+renderMissingZettelLink zid = do
+  let classes = ["zettel-link-container", "errors"]
+  elClass "span" (T.intercalate " " classes) $ do
+    let errMsg = "Wiki-link does not refer to any existing zettel"
+    elAttr "span" ("class" =: "zettel-link" <> "title" =: errMsg) $ do
+      elAttr "a" mempty $ text $ unZettelID zid
+
+-- | Like `renderZettelLink` but when we only have ID in hand.
+renderZettelLinkIDOnly :: (DomBuilder t m, PostBuild t m) => ZettelID -> Slug -> NeuronWebT t m ()
+renderZettelLinkIDOnly zid slug =
+  elClass "span" "zettel-link-container" $ do
+    elClass "span" "zettel-link" $ do
+      neuronRouteLink (Some $ Route_Zettel slug) mempty $ text $ unZettelID zid
+
+zettelLinkCss :: Css
+zettelLinkCss = do
+  "span.zettel-link-container span.zettel-link a" ? do
+    C.fontWeight C.bold
+    C.textDecoration C.none
+  "span.zettel-link-container span.extra" ? do
+    C.color C.auto
+  "span.zettel-link-container.errors" ? do
+    C.border C.solid (C.px 1) C.red
+  "[data-tooltip]:after" ? do
+    C.fontSize $ em 0.7
 
 -- Markdown parsing
 -- ----------------
@@ -226,15 +303,3 @@ wikiLinkSpec =
         -- Store connetion type in 'title' attribute
         -- TODO: Put it in attrs instead; requires PR to commonmark
         title = show conn
-
-zettelLinkCss :: Css
-zettelLinkCss = do
-  "span.zettel-link-container span.zettel-link a" ? do
-    C.fontWeight C.bold
-    C.textDecoration C.none
-  "span.zettel-link-container span.extra" ? do
-    C.color C.auto
-  "span.zettel-link-container.errors" ? do
-    C.border C.solid (C.px 1) C.red
-  "[data-tooltip]:after" ? do
-    C.fontSize $ em 0.7
